- Created debug build extending default build. - Moved Stetho into debug build.
