+++ app/src/main/java/org/schabi/newpipe/App.java
@@ -66,7 +66,6 @@ public class App extends Application {
         }
 
         NewPipeDatabase.getInstance( getApplicationContext() );
-        initStetho();
 
         //init NewPipe
         NewPipe.init(Downloader.getInstance());
@@ -112,27 +111,5 @@ public class App extends Application {
 
     public static boolean isUsingTor() {
         return useTor;
-    }
-
-    private void initStetho() {
-        // Create an InitializerBuilder
-        Stetho.InitializerBuilder initializerBuilder =
-                Stetho.newInitializerBuilder(this);
-
-        // Enable Chrome DevTools
-        initializerBuilder.enableWebKitInspector(
-                Stetho.defaultInspectorModulesProvider(this)
-        );
-
-        // Enable command line interface
-        initializerBuilder.enableDumpapp(
-                Stetho.defaultDumperPluginsProvider(getApplicationContext())
-        );
-
-        // Use the InitializerBuilder to generate an Initializer
-        Stetho.Initializer initializer = initializerBuilder.build();
-
-        // Initialize Stetho with the Initializer
-        Stetho.initialize(initializer);
     }
 }
